[gac] refactor: rename Employee to User
