+++ src/main/java/scrumbackend/team/Team.java
@@ -11,7 +11,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import scrumbackend.employee.Employee;
+import scrumbackend.user.User;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -28,5 +28,5 @@ public class Team {
   private String name;
 
   @OneToMany(mappedBy = "team")
-  private List<Employee> employees;
+  private List<User> users;
 }
